+++ legacy/gff/gff_byn_7/source/help/gff_byn_7.php
@@ -114,7 +114,7 @@ See <a target="_blank" href='BlinTyping-English.pdf'>Typing in Blin</a> for full
 <h2>Troubleshooting</h2>
 <p class='keymanweb'>If the characters you type or those on the On Screen Keyboard do not appear to display correctly, please read the <a target="_blank" href="/troubleshooting/#boxes">KeymanWeb troubleshooting guide</a>.</p>
 
-<p>For any other questions, <a target="_blank" href="http://www.tavultesoft.com/contact.php">contact us</a>.</p>
+<p>For any other questions, <a target="_blank" href="https://keyman.com/contact/">contact us</a>.</p>
 
 </div>
 
@@ -139,10 +139,10 @@ See <a target="_blank" href='BlinTyping-English.pdf'>Typing in Blin</a> for full
 </div>
 <div>
 <h3>Other Internet Resources</h3>
-<p><a target="_blank" href="http://keymankeyboards.com/?id=458">Download the Blin Keyboard for Keyman Desktop</a></p>
-<p><a target="_blank" href="http://en.wikipedia.org/wiki/Blin_language">More on the Blin Language at Wikipedia.org</a></p>
+<p><a target="_blank" href="https://keyman.com/keyboards/gff-byn-powerpack-7">Download the Blin Keyboard for Keyman Desktop</a></p>
+<p><a target="_blank" href="https://en.wikipedia.org/wiki/Bilen_language">More on the Blin Language at Wikipedia.org</a></p>
 <p><a target="_blank" href="http://www.omniglot.com/writing/ethiopic.htm">More on the Ge'ez Script at Omniglot.com</a><br/>
-<a target="_blank" href="http://en.wikipedia.org/wiki/Ge'ez_alphabet">More on the Ge'ez Script at Wikipedia.org</a></p>
+<a target="_blank" href="https://en.wikipedia.org/wiki/Ge%27ez_script">More on the Ge'ez Script at Wikipedia.org</a></p>
 </div>
 </div>