+++ legacy/gff/gff_mym_7/source/help/gff_mym_7.php
@@ -120,7 +120,7 @@ See <a target="_blank" href='MeenTyping-English.pdf'>Typing in <?php echo $langN
 <h2>Troubleshooting</h2>
 <p class='keymanweb'>If the characters you type or those on the On Screen Keyboard do not appear to display correctly, please read the <a target="_blank" href="/troubleshooting/#boxes">KeymanWeb troubleshooting guide</a>.</p>
 
-<p>For any other questions, <a target="_blank" href="http://www.tavultesoft.com/contact.php">contact us</a>.</p>
+<p>For any other questions, <a target="_blank" href="https://keyman.com/contact/">contact us</a>.</p>
 
 </div>
 
@@ -146,10 +146,10 @@ See <a target="_blank" href='MeenTyping-English.pdf'>Typing in <?php echo $langN
 </div>
 <div>
 <h3>Other Internet Resources</h3>
-<p><a target="_blank" href="http://keymankeyboards.com/?id=462">Download the <?php echo $langName; ?> Keyboard for Keyman Desktop</a></p>
+<p><a target="_blank" href="https://keyman.com/keyboards/gff-mym-powerpack-7">Download the <?php echo $langName; ?> Keyboard for Keyman Desktop</a></p>
 <p><a target="_blank" href="http://en.wikipedia.org/wiki/<?php if($lang == 'suq') echo "Surma_people"; else echo "{$langName}_language"; ?>">More on the <?php echo $langName; ?> Language at Wikipedia.org</a></p>
 <p><a target="_blank" href="http://www.omniglot.com/writing/ethiopic.htm">More on the Ge'ez Script at Omniglot.com</a><br/>
-<a target="_blank" href="http://en.wikipedia.org/wiki/Ge'ez_alphabet">More on the Ge'ez Script at Wikipedia.org</a></p>
+<a target="_blank" href="https://en.wikipedia.org/wiki/Ge%27ez_script">More on the Ge'ez Script at Wikipedia.org</a></p>
 </div>
 </div>