+++ legacy/gff/gff_sgw_7/source/help/gff_sgw_7.php
@@ -118,7 +118,7 @@ for full details on how to type all Sebatbeit letters, numbers and punctuation.
 <h2>Troubleshooting</h2>
 <p class='keymanweb'>If the characters you type or those on the On Screen Keyboard do not appear to display correctly, please read the <a target="_blank" href="/troubleshooting/#boxes">KeymanWeb troubleshooting guide</a>.</p>
 
-<p>For any other questions, <a target="_blank" href="http://www.tavultesoft.com/contact.php">contact us</a>.</p>
+<p>For any other questions, <a target="_blank" href="https://keyman.com/contact/">contact us</a>.</p>
 
 </div>
 
@@ -144,10 +144,10 @@ for full details on how to type all Sebatbeit letters, numbers and punctuation.
 </div>
 <div>
 <h3>Other Internet Resources</h3>
-<p><a target="_blank" href="http://keymankeyboards.com/?id=463">Download the Sebatbeit Keyboard for Keyman Desktop</a></p>
-<a target="_blank" href="http://en.wikipedia.org/wiki/Sebat_bet_gurage">More on the Sebatbeit Language at Wikipedia.org</a><br/>
+<p><a target="_blank" href="https://keyman.com/keyboards/gff-sgw-powerpack-7">Download the Sebatbeit Keyboard for Keyman Desktop</a></p>
+<a target="_blank" href="https://en.wikipedia.org/wiki/Sebat_Bet_Gurage_language">More on the Sebatbeit Language at Wikipedia.org</a><br/>
 <p><a target="_blank" href="http://www.omniglot.com/writing/ethiopic.htm">More on the Ge'ez Script at Omniglot.com</a><br/>
-<a target="_blank" href="http://en.wikipedia.org/wiki/Ge'ez_alphabet">More on the Ge'ez Script at Wikipedia.org</a></p>
+<a target="_blank" href="https://en.wikipedia.org/wiki/Ge%27ez_script">More on the Ge'ez Script at Wikipedia.org</a></p>
 </div>
 </div>