+++ legacy/n/ndailu keyman keyboard/source/help/ndailu keyman keyboard.php
@@ -1,0 +1,31 @@
+﻿<?php
+  $pagename = 'DaiLue Keyboard Help';
+  $pagetitle = 'DaiLue Keyboard Help';
+  require_once('header.php');
+?>
+
+<p>This keyboard layout is designed for DaiLue.  It includes an on screen keyboard
+which can be viewed by clicking on the Keyman icon and selecting the On Screen Keyboard
+menu item.<p>  
+
+<p>DiaLue keyboard<p>
+<h2><img src="DaiLue Keyboard.png" alt="" height="198" width="514" border="0" /></h2
+<p>DiaLue keyboard in Shifted state<p>
+<h2><img src="DaiLue Keyboard Shifted.png" alt="" height="198" width="514" border="0" /></h2
+
+
+<p>You can type an additional 6 characters by either double tapping or triple tapping certain keys.<p>
+
+<p>Characters typed by double tapping the same key
+<p>Typing ‘g’ twice will present the character ᦶ (U+19B6)<p>
+<p>Typing ‘W’ twice will present the character “ (U+201C)<p>
+<p>Typing ‘E’ twice will present the character ” (U+201D)<p>
+<p>Typing ‘.’ twice will present the character 、 (U+3001)<p>
+<p>Typing ‘_’ twice will present you with – (U+2013)<p>
+<p>Character typed by triple tapping the same key<p>
+<p>Typing ‘_’ three times will present you with — (U+2014)<p>  
+
+
+
+<p>SIL's Dai Banna SIL font is recommended for use with DaiLue and can be found on by visiting Scriptsource.org
+and searching for Dai Banna Fonts</p>