+++ release/sil/sil_bari/source/help/sil_bari.php
@@ -1,0 +1,8 @@
+<?php
+  $pagename = 'Bari (SIL) Keyboard Help';
+  $pagetitle = 'Bari (SIL) Keyboard Help';
+  require_once('header.php');
+?>
+    <h1>Bari (SIL) Keyboard</h1>
+    <p>Bari (SIL) keyboard for the Bari language of South Sudan.</p>
+<img src='keyboard_layout.png' alt='Keyboard Layout' width='720px'>