+++ release/sil/sil_dzongkha/source/help/sil_dzongkha.php
@@ -1,0 +1,167 @@
+<?php
+  $pagename = 'Dzongkha (SIL)';
+  $pagetitle = 'Dzongkha (SIL)';
+	$pagestyle = <<<END
+	samp{font-family: 'DDC Uchen', 'Tibetan Machine Uni', 'Microsoft Himalaya'; color:blue;}
+	kbd {color:black;}
+	table {border-collapse:collapse !important; margin:10px 0px 20px 40px;}
+	.grid col {background-color:#ffffff;}
+	.grid td, .grid th {padding:2px 5px;border-bottom:1px solid #ddd;}
+	.grid tr {text-align:center; font-weight:normal; height:80px;}
+	.grid thead tr, .grid .headrow{height:30px;}
+	.grid .headrow {font-weight:bold; color:#AD4A28; background-color:#eee; text-align:left;}
+	col.char   {width:85px;}
+	col.keys   {width:75px;}
+	col.point {width:85px;}
+	col.name   {width:165px;}
+	td samp {font-size:24pt}
+	table kbd {font-size:12pt;padding:2px 4px; vertical-align:middle;}
+	.columns  {width:375px; -webkit-column-count: 2; -webkit-column-gap: 10px; -moz-column-count: 2; -moz-column-gap: 10px; column-count: 2; column-gap: 10px;}
+	.colleft, .colright {position:relative; float:left;}
+	.colright {padding-left:40px;}
+	p samp {font-size:22pt; line-height:.8em;}
+	p kbd {color:black; font:9pt Arial; border:solid 1px grey; background:#ccc; margin:0px 1px; padding:0px 3px; -webkit-border-radius: 3px; -moz-border-radius: 3px; border-radius: 3px;}
+					.key {
+							float: left;
+							display: block;
+							position: relative;
+							overflow: hidden;
+							height: 35px;
+							margin: 2px 0px 0px 2px;
+							border: solid 1px grey;
+				-webkit-border-radius: 3px; -moz-border-radius: 3px; border-radius: 3px;
+					}
+					.plain {
+						background-repeat: no-repeat;
+						width: 34px;
+					}
+			.empty {
+				background:#eee;
+			}
+			.K_BKQUOTE {
+			width:34px;
+			}
+					.K_BKSP {
+			width: 60px;
+			}
+			.K_TAB {
+			width: 52px;
+			}
+					.K_BKSLASH {
+					width: 42px;
+					}
+			.K_CAPS {
+			width: 62px;
+			}
+			.K_ENTER {
+					width: 70px;
+			}
+			.K_ENTER-EU1 {
+					width: 42px;
+			-webkit-border-radius: 3px 3px 0px; -moz-border-radius: 3px 3px 0px; border-radius: 3px 3px 0px;
+			}
+			.K_ENTER-EU1 .key kbd {
+			left:3px !important;
+			}
+			.K_ENTER-EU2 {
+					width: 32px;
+			margin-top:-1px !important;
+			border-top:none !important;
+			height:39px !important;
+			-webkit-border-radius: 0px 0px 3px 3px; -moz-border-radius: 0px 0px 3px 3px; border-radius: 0px 0px 3px 3px;
+			}
+					.K_SHIFTL {
+						width: 85px;
+					}
+					.K_SHIFTL-EU {
+						width: 47px;
+					}
+					.K_oE2 {
+						display:none;
+					}
+			.K_SHIFTR {
+						width: 85px;
+					}
+			.K_CTRLL {
+				width:60px;
+				margin-right: 40px;
+					}
+			.K_ALTL {
+						width: 40px;
+					}
+					.K_SPACE , .K_SPACE samp{
+						width: 224px;
+					}
+			.K_ALTR {
+						width: 40px;
+					}
+			.K_CTRLR {
+						width: 60px;
+				margin-left: 76px;
+			}
+
+
+					.special {
+						background-repeat: no-repeat;
+				background: #c0c0c0;
+					}
+
+					.key kbd {
+						font: bold 7pt Arial;
+						position: absolute;
+						left: 3px;
+						top: 3px;
+					}
+
+					.key kbd {
+						display: block;
+					}
+
+					.special kbd {
+						display: block;
+					}
+
+					.key samp {
+				font-size:20px;
+						position: absolute;
+						display: block;
+						right: 3px;
+						bottom: -3px;
+					}
+			#altgr .K_J samp, #shift_altgr .K_J samp {
+			bottom:3px;
+			}
+			.K_SPACE samp {
+			 text-align:center;
+				}
+
+			.LTN {
+				font-family:Arial !important;
+				font-size:12px !important;
+				color: #AD4A28 !important;
+				right:2px !important;
+				bottom:2px !important;
+				font-variant:small-caps;
+			}
+
+END;
+  require_once('header.php');
+?>
+
+
+
+
+
+
+
+
+	<h2>Overview</h2>
+	<p>This keyboard is designed for the Dzongkha language of Bhutan. The keyboard follows the current (October 2009) official Dzongkha layout approved by the Department of Information Technology and Dzongkha Development Authority of the Government of Bhutan. This keyboard is fully compliant with the <a href='http://unicode.org/standard/standard.html'>Unicode Standard</a>.</p>	<h4>Basic Function</h4>	<p>This keyboard has four states: unshifted, Shift, AltGr, and Shift AltGr. The AltGr state can also be called Right Alt. If your hardware keyboard doesn't have an <kbd>AltGr</kbd> key, you can simulate it. From the Options tab of Keyman Configuration, tick 'Simulate AltGr with Ctrl+Alt' and click <button type='button'>OK</button>. You can then use <kbd>Ctrl</kbd>+<kbd>Alt</kbd> to type any keys in the AltGr state and <kbd>Shift</kbd>+<kbd>Ctrl</kbd>+<kbd>Alt</kbd> to type any keys in the Shift AltGr state.</p>	<p>You can also type Latin letters while using this keyboard. To get the standard Latin keystroke for any key instead of the Dzongkha output, type <kbd>$</kbd> first. For example, to get <samp>%</samp> instead of <samp>༎</samp> type <kbd>$</kbd><kbd>%</kbd>.
+	<h4>On Screen Keyboard</h4>
+	<p>This keyboard includes an On Screen Keyboard view for easy reference. The On Screen Keyboard works best when associated with a QWERTY US layout.</p>
+	<h4>Fonts</h4>
+	<p>This is a Unicode keyboard and works with any Unicode font which has support for Tibetan script. The <b>DDC Uchen</b> font fully supports this keyboard and is installed with the keyboard. To find other fonts on your computer which support the keyboard, use the Keyman <a href='http://help.keymandesktop.com/docs/start_font.php'>Font Helper</a>.</p>
+  </section>
+  <section id='layout'>
+	<h2>Keyboard Layout</h2>	<article id='unshifted'>		<h3 style="clear:left">Unshifted</h3>		<div class="key plain K_BKQUOTE">			<kbd>`</kbd>			<samp>༉</samp>		</div>		<div class="key plain K_1">			<kbd>1</kbd>			<samp>༡</samp>		</div>		<div class="key plain K_2">			<kbd>2</kbd>			<samp>༢</samp>		</div>		<div class="key plain K_3">			<kbd>3</kbd>			<samp>༣</samp>		</div>		<div class="key plain K_4">			<kbd>4</kbd>			<samp>༤</samp>		</div>		<div class="key plain K_5">			<kbd>5</kbd>			<samp>༥</samp>		</div>		<div class="key plain K_6">			<kbd>6</kbd>			<samp>༦</samp>		</div>		<div class="key plain K_7">			<kbd>7</kbd>			<samp>༧</samp>		</div>		<div class="key plain K_8">			<kbd>8</kbd>			<samp>༨</samp>		</div>		<div class="key plain K_9">			<kbd>9</kbd>			<samp>༩</samp>		</div>		<div class="key plain K_0">			<kbd>0</kbd>			<samp>༠</samp>		</div>		<div class="key plain K_HYPHEN">			<kbd>-</kbd>			<samp>༔</samp>		</div>		<div class="key plain K_EQUAL">			<kbd>=</kbd>			<samp>།</samp>		</div>		<div class="key special K_BKSP">			<kbd>Backspace</kbd>		</div>		<br style="clear: left">		<div class="key special K_TAB">			<kbd>Tab</kbd>		</div>		<div class="key plain K_Q">			<kbd>q</kbd>			<samp>ཀ</samp>		</div>		<div class="key plain K_W">			<kbd>w</kbd>			<samp>ཁ</samp>		</div>		<div class="key plain K_E">			<kbd>e</kbd>			<samp>ག</samp>		</div>		<div class="key plain K_R">			<kbd>r</kbd>			<samp>ང</samp>		</div>		<div class="key plain K_T">			<kbd>t</kbd>			<samp>ི</samp>		</div>		<div class="key plain K_Y">			<kbd>y</kbd>			<samp>ུ</samp>		</div>		<div class="key plain K_U">			<kbd>u</kbd>			<samp>ེ</samp>		</div>		<div class="key plain K_I">			<kbd>i</kbd>			<samp>ོ</samp>		</div>		<div class="key plain K_O">			<kbd>o</kbd>			<samp>ཅ</samp>		</div>		<div class="key plain K_P">			<kbd>p</kbd>			<samp>ཆ</samp>		</div>		<div class="key plain K_LBRKT">			<kbd>[</kbd>			<samp>ཇ</samp>		</div>		<div class="key plain K_RBRKT">			<kbd>]</kbd>			<samp>ཉ</samp>		</div>		<div class="key plain K_BKSLASH">			<kbd>\</kbd>			<samp>ཝ</samp>		</div>		<br style="clear: left">		<div class="key special K_CAPS">			<kbd>Caps Lock</kbd>		</div>		<div class="key plain K_A">			<kbd>a</kbd>			<samp>ཏ</samp>		</div>		<div class="key plain K_S">			<kbd>s</kbd>			<samp>ཐ</samp>		</div>		<div class="key plain K_D">			<kbd>d</kbd>			<samp>ད</samp>		</div>		<div class="key plain K_F">			<kbd>f</kbd>			<samp>ན</samp>		</div>		<div class="key plain K_G">			<kbd>g</kbd>			<samp>པ</samp>		</div>		<div class="key plain K_H">			<kbd>h</kbd>			<samp>ཕ</samp>		</div>		<div class="key plain K_J">			<kbd>j</kbd>			<samp>བ</samp>		</div>		<div class="key plain K_K">			<kbd>k</kbd>			<samp>མ</samp>		</div>		<div class="key plain K_L">			<kbd>l</kbd>			<samp>ཙ</samp>		</div>		<div class="key plain K_COLON">			<kbd>;</kbd>			<samp>ཚ</samp>		</div>		<div class="key plain K_QUOTE">			<kbd>'</kbd>			<samp>ཛ</samp>		</div>		<div class="key special K_ENTER">			<kbd>Enter</kbd>		</div>		<br style="clear: left">		<div class="key special K_SHIFTL">			<kbd>Shift</kbd>		</div>		<div class="key plain K_oE2">			<kbd>\</kbd>		</div>		<div class="key plain K_Z">			<kbd>z</kbd>			<samp>ཞ</samp>		</div>		<div class="key plain K_X">			<kbd>x</kbd>			<samp>ཟ</samp>		</div>		<div class="key plain K_C">			<kbd>c</kbd>			<samp>འ</samp>		</div>		<div class="key plain K_V">			<kbd>v</kbd>			<samp>ཡ</samp>		</div>		<div class="key plain K_B">			<kbd>b</kbd>			<samp>ར</samp>		</div>		<div class="key plain K_N">			<kbd>n</kbd>			<samp>ལ</samp>		</div>		<div class="key plain K_M">			<kbd>m</kbd>			<samp>ཤ</samp>		</div>		<div class="key plain K_COMMA">			<kbd>,</kbd>			<samp>ས</samp>		</div>		<div class="key plain K_PERIOD">			<kbd>.</kbd>			<samp>ཧ</samp>		</div>		<div class="key plain K_SLASH">			<kbd>/</kbd>			<samp>ཨ</samp>		</div>		<div class="key special K_SHIFTR">			<kbd>Shift</kbd>		</div>		<br style="clear: left">		<div class="key special K_CTRLL">			<kbd>Ctrl</kbd>		</div>		<div class="key special K_ALTL">			<kbd>Alt</kbd>		</div>		<div class="key plain K_SPACE">			<kbd> </kbd>			<samp>༌<span class='LTN'>‹tsheg›</span></samp>		</div>		<div class="key special K_ALTR">			<kbd>Alt</kbd>		</div>		<div class="key special K_CTRLR">			<kbd>Ctrl</kbd>		</div>	</article>	<article id='shift' class='break'>		<h3 style="clear:left">Shift</h3>		<div class="key plain K_BKQUOTE">			<kbd>~</kbd>			<samp>༊</samp>		</div>		<div class="key plain K_1">			<kbd>!</kbd>			<samp>༄</samp>		</div>		<div class="key plain K_2">			<kbd>@</kbd>			<samp>༅</samp>		</div>		<div class="key plain K_3">			<kbd>#</kbd>			<samp>༆</samp>		</div>		<div class="key plain K_4">			<kbd>$</kbd>			<samp class="LTN">‹LTN›</samp>		</div>		<div class="key plain K_5">			<kbd>%</kbd>			<samp>༎</samp>		</div>		<div class="key plain K_6">			<kbd>^</kbd>			<samp>༈</samp>		</div>		<div class="key plain K_7">			<kbd>&amp;</kbd>			<samp>༸</samp>		</div>		<div class="key plain K_8">			<kbd>*</kbd>			<samp>༴</samp>		</div>		<div class="key plain K_9">			<kbd>(</kbd>			<samp>༼</samp>		</div>		<div class="key plain K_0">			<kbd>)</kbd>			<samp>༽</samp>		</div>		<div class="key plain K_HYPHEN">			<kbd>_</kbd>			<samp>ཿ</samp>		</div>		<div class="key plain K_EQUAL">			<kbd>+</kbd>			<samp>༑</samp>		</div>		<div class="key special K_BKSP">			<kbd>Backspace</kbd>		</div>		<br style="clear: left">		<div class="key special K_TAB">			<kbd>Tab</kbd>		</div>		<div class="key plain K_Q">			<kbd>Q</kbd>			<samp>ྐ</samp>		</div>		<div class="key plain K_W">			<kbd>W</kbd>			<samp>ྑ</samp>		</div>		<div class="key plain K_E">			<kbd>E</kbd>			<samp>ྒ</samp>		</div>		<div class="key plain K_R">			<kbd>R</kbd>			<samp>ྔ</samp>		</div>		<div class="key plain K_T">			<kbd>T</kbd>			<samp>ྀ</samp>		</div>		<div class="key plain K_Y">			<kbd>Y</kbd>			<samp>྄</samp>		</div>		<div class="key plain K_U">			<kbd>U</kbd>			<samp>ཻ</samp>		</div>		<div class="key plain K_I">			<kbd>I</kbd>			<samp>ཽ</samp>		</div>		<div class="key plain K_O">			<kbd>O</kbd>			<samp>ྕ</samp>		</div>		<div class="key plain K_P">			<kbd>P</kbd>			<samp>ྖ</samp>		</div>		<div class="key plain K_LBRKT">			<kbd>{</kbd>			<samp>ྗ</samp>		</div>		<div class="key plain K_RBRKT">			<kbd>}</kbd>			<samp>ྙ</samp>		</div>		<div class="key plain K_BKSLASH">			<kbd>|</kbd>			<samp>ྭ</samp>		</div>		<br style="clear: left">		<div class="key special K_CAPS">			<kbd>Caps Lock</kbd>		</div>		<div class="key plain K_A">			<kbd>A</kbd>			<samp>ྟ</samp>		</div>		<div class="key plain K_S">			<kbd>S</kbd>			<samp>ྠ</samp>		</div>		<div class="key plain K_D">			<kbd>D</kbd>			<samp>ྡ</samp>		</div>		<div class="key plain K_F">			<kbd>F</kbd>			<samp>ྣ</samp>		</div>		<div class="key plain K_G">			<kbd>G</kbd>			<samp>ྤ</samp>		</div>		<div class="key plain K_H">			<kbd>H</kbd>			<samp>ྥ</samp>		</div>		<div class="key plain K_J">			<kbd>J</kbd>			<samp>ྦ</samp>		</div>		<div class="key plain K_K">			<kbd>K</kbd>			<samp>ྨ</samp>		</div>		<div class="key plain K_L">			<kbd>L</kbd>			<samp>ྩ</samp>		</div>		<div class="key plain K_COLON">			<kbd>:</kbd>			<samp>ྪ</samp>		</div>		<div class="key plain K_QUOTE">			<kbd>"</kbd>			<samp>ྫ</samp>		</div>		<div class="key special K_ENTER">			<kbd>Enter</kbd>		</div>		<br style="clear: left">		<div class="key special K_SHIFTL">			<kbd>Shift</kbd>		</div>		<div class="key plain K_oE2">			<kbd>\</kbd>		</div>		<div class="key plain K_Z">			<kbd>Z</kbd>			<samp>ྮ</samp>		</div>		<div class="key plain K_X">			<kbd>X</kbd>			<samp>ྯ</samp>		</div>		<div class="key plain K_C">			<kbd>C</kbd>			<samp>ཱ</samp>		</div>		<div class="key plain K_V">			<kbd>V</kbd>			<samp>ྱ</samp>		</div>		<div class="key plain K_B">			<kbd>B</kbd>			<samp>ྲ</samp>		</div>		<div class="key plain K_N">			<kbd>N</kbd>			<samp>ླ</samp>		</div>		<div class="key plain K_M">			<kbd>M</kbd>			<samp>ྴ</samp>		</div>		<div class="key plain K_COMMA">			<kbd>&lt;</kbd>			<samp>ྶ</samp>		</div>		<div class="key plain K_PERIOD">			<kbd>&gt;</kbd>			<samp>ྷ</samp>		</div>		<div class="key plain K_SLASH">			<kbd>?</kbd>			<samp>ྸ</samp>		</div>		<div class="key special K_SHIFTR">			<kbd>Shift</kbd>		</div>		<br style="clear: left">		<div class="key special K_CTRLL">			<kbd>Ctrl</kbd>		</div>		<div class="key special K_ALTL">			<kbd>Alt</kbd>		</div>		<div class="key plain K_SPACE">			<kbd> </kbd>			<samp class='LTN'>‹space›</samp>		</div>		<div class="key special K_ALTR">			<kbd>Alt</kbd>		</div>		<div class="key special K_CTRLR">			<kbd>Ctrl</kbd>		</div>	</article>	<article id='altgr'>		<h3 style="clear:left">AltGr</h3>		<div class="key plain ctrl K_BKQUOTE">			<kbd>`</kbd>			<samp>࿑</samp>		</div>		<div class="key plain K_1">			<kbd>1</kbd>			<samp>1</samp>		</div>		<div class="key plain K_2">			<kbd>2</kbd>			<samp>2</samp>		</div>		<div class="key plain K_3">			<kbd>3</kbd>			<samp>3</samp>		</div>		<div class="key plain K_4">			<kbd>4</kbd>			<samp>4</samp>		</div>		<div class="key plain K_5">			<kbd>5</kbd>			<samp>5</samp>		</div>		<div class="key plain K_6">			<kbd>6</kbd>			<samp>6</samp>		</div>		<div class="key plain K_7">			<kbd>7</kbd>			<samp>7</samp>		</div>		<div class="key plain K_8">			<kbd>8</kbd>			<samp>8</samp>		</div>		<div class="key plain K_9">			<kbd>9</kbd>			<samp>9</samp>		</div>		<div class="key plain K_0">			<kbd>0</kbd>			<samp>0</samp>		</div>		<div class="key plain K_HYPHEN">			<kbd>-</kbd>			<samp>-</samp>		</div>		<div class="key plain K_EQUAL">			<kbd>=</kbd>			<samp>=</samp>		</div>		<div class="key special K_BKSP">			<kbd>Backspace</kbd>		</div>		<br style="clear: left">		<div class="key special K_TAB">			<kbd>Tab</kbd>		</div>		<div class="key plain K_Q">			<kbd>q</kbd>			<samp>ྈ</samp>		</div>		<div class="key plain K_W">			<kbd>w</kbd>			<samp>ྉ</samp>		</div>		<div class="key plain K_E">			<kbd>e</kbd>			<samp>ྌ</samp>		</div>		<div class="key plain K_R">			<kbd>r</kbd>			<samp>ྃ</samp>		</div>		<div class="key plain K_T">			<kbd>t</kbd>			<samp>༚</samp>		</div>		<div class="key plain K_Y">			<kbd>y</kbd>			<samp>༛</samp>		</div>		<div class="key plain K_U">			<kbd>u</kbd>			<samp>༜</samp>		</div>		<div class="key plain K_I">			<kbd>i</kbd>			<samp>༝</samp>		</div>		<div class="key plain K_O">			<kbd>o</kbd>			<samp>༞</samp>		</div>		<div class="key plain K_P">			<kbd>p</kbd>			<samp>༟</samp>		</div>		<div class="key plain K_LBRKT">			<kbd>[</kbd>			<samp>(</samp>		</div>		<div class="key plain K_RBRKT">			<kbd>]</kbd>			<samp>)</samp>		</div>		<div class="key plain K_BKSLASH">			<kbd>\</kbd>			<samp>\</samp>		</div>		<br style="clear: left">		<div class="key special K_CAPS">			<kbd>Caps Lock</kbd>		</div>		<div class="key plain K_A">			<kbd>a</kbd>			<samp>ཊ</samp>		</div>		<div class="key plain K_S">			<kbd>s</kbd>			<samp>ཋ</samp>		</div>		<div class="key plain K_D">			<kbd>d</kbd>			<samp>ཌ</samp>		</div>		<div class="key plain K_F">			<kbd>f</kbd>			<samp>ཎ</samp>		</div>		<div class="key plain K_G empty">			<kbd>g</kbd>			<samp></samp>		</div>		<div class="key plain K_H empty">			<kbd>h</kbd>			<samp></samp>		</div>		<div class="key plain K_J">			<kbd>j</kbd>			<samp>༷</samp>		</div>		<div class="key plain K_K">			<kbd>k</kbd>			<samp>ཾ</samp>		</div>		<div class="key plain K_L">			<kbd>l</kbd>			<samp>༹</samp>		</div>		<div class="key plain K_COLON">			<kbd>;</kbd>			<samp>;</samp>		</div>		<div class="key plain K_QUOTE">			<kbd>'</kbd>			<samp>'</samp>		</div>		<div class="key special K_ENTER">			<kbd>Enter</kbd>		</div>		<br style="clear: left">		<div class="key special K_SHIFTL">			<kbd>Shift</kbd>		</div>		<div class="key plain K_oE2">			<kbd>\</kbd>		</div>		<div class="key plain K_Z">			<kbd>z</kbd>			<samp>༓</samp>		</div>		<div class="key plain K_X">			<kbd>x</kbd>			<samp>྾</samp>		</div>		<div class="key plain K_C">			<kbd>c</kbd>			<samp>༃</samp>		</div>		<div class="key plain K_V">			<kbd>v</kbd>			<samp>༏</samp>		</div>		<div class="key plain K_B">			<kbd>b</kbd>			<samp>ཪ</samp>		</div>		<div class="key plain K_N">			<kbd>n</kbd>			<samp>༒</samp>		</div>		<div class="key plain K_M">			<kbd>m</kbd>			<samp>ཥ</samp>		</div>		<div class="key plain K_COMMA">			<kbd>,</kbd>			<samp>,</samp>		</div>		<div class="key plain K_PERIOD">			<kbd>.</kbd>			<samp>.</samp>		</div>		<div class="key plain K_SLASH">			<kbd>/</kbd>			<samp>/</samp>		</div>		<div class="key special K_SHIFTR">			<kbd>Shift</kbd>		</div>		<br style="clear: left">		<div class="key special K_CTRLL">			<kbd>Ctrl</kbd>		</div>		<div class="key special K_ALTL">			<kbd>Alt</kbd>		</div>		<div class="key plain K_SPACE">			<kbd> </kbd>			<samp>༌<span class='LTN'>‹no-break tsheg›</span></samp>		</div>		<div class="key special K_ALTR">			<kbd>Alt</kbd>		</div>		<div class="key special K_CTRLR">			<kbd>Ctrl</kbd>		</div>	</article>	<article id='shift_altgr'>		<h3 style="clear:left">Shift AltGr</h3>		<div class="key plain K_BKQUOTE">			<kbd>~</kbd>			<samp>࿐</samp>		</div>		<div class="key plain K_1">			<kbd>!</kbd>			<samp>࿓</samp>		</div>		<div class="key plain K_2">			<kbd>@</kbd>			<samp>࿔</samp>		</div>		<div class="key plain K_3">			<kbd>#</kbd>			<samp>༺</samp>		</div>		<div class="key plain K_4">			<kbd>$</kbd>			<samp>༻</samp>		</div>		<div class="key plain K_5">			<kbd>%</kbd>			<samp>྅</samp>		</div>		<div class="key plain K_6">			<kbd>^</kbd>			<samp>༁</samp>		</div>		<div class="key plain K_7">			<kbd>&amp;</kbd>			<samp>ྊ</samp>		</div>		<div class="key plain K_8">			<kbd>*</kbd>			<samp>*</samp>		</div>		<div class="key plain K_9 empty">			<kbd>(</kbd>			<samp></samp>		</div>		<div class="key plain K_0 empty">			<kbd>)</kbd>			<samp></samp>		</div>		<div class="key plain K_HYPHEN">			<kbd>_</kbd>			<samp>࿒</samp>		</div>		<div class="key plain K_EQUAL">			<kbd>+</kbd>			<samp>+</samp>		</div>		<div class="key special K_BKSP">			<kbd>Backspace</kbd>		</div>		<br style="clear: left">		<div class="key special K_TAB">			<kbd>Tab</kbd>		</div>		<div class="key plain K_Q">			<kbd>Q</kbd>			<samp> ྍ</samp>		</div>		<div class="key plain K_W">			<kbd>W</kbd>			<samp> ྎ</samp>		</div>		<div class="key plain K_E">			<kbd>E</kbd>			<samp> ྏ</samp>		</div>		<div class="key plain K_R">			<kbd>R</kbd>			<samp>ྂ</samp>		</div>		<div class="key plain K_T">			<kbd>T</kbd>			<samp>ྋྙ</samp>		</div>		<div class="key plain K_Y empty">			<kbd>Y</kbd>			<samp></samp>		</div>		<div class="key plain K_U empty">			<kbd>U</kbd>			<samp></samp>		</div>		<div class="key plain K_I">			<kbd>I</kbd>			<samp>༗</samp>		</div>		<div class="key plain K_O">			<kbd>O</kbd>			<samp>༘</samp>		</div>		<div class="key plain K_P">			<kbd>P</kbd>			<samp>༙</samp>		</div>		<div class="key plain K_LBRKT">			<kbd>{</kbd>			<samp>༿</samp>		</div>		<div class="key plain K_RBRKT">			<kbd>}</kbd>			<samp>༾</samp>		</div>		<div class="key plain K_BKSLASH">			<kbd>|</kbd>			<samp>ྺ</samp>		</div>		<br style="clear: left">		<div class="key special K_CAPS">			<kbd>Caps Lock</kbd>		</div>		<div class="key plain K_A">			<kbd>A</kbd>			<samp>ྚ</samp>		</div>		<div class="key plain K_S">			<kbd>S</kbd>			<samp>ྛ</samp>		</div>		<div class="key plain K_D">			<kbd>D</kbd>			<samp>ྜ</samp>		</div>		<div class="key plain K_F">			<kbd>F</kbd>			<samp>ྞ</samp>		</div>		<div class="key plain K_G empty">			<kbd>G</kbd>			<samp></samp>		</div>		<div class="key plain K_H empty">			<kbd>H</kbd>			<samp></samp>		</div>		<div class="key plain K_J">			<kbd>J</kbd>			<samp>༵</samp>		</div>		<div class="key plain K_K">			<kbd>K</kbd>			<samp>྆</samp>		</div>		<div class="key plain K_L">			<kbd>L</kbd>			<samp>྇</samp>		</div>		<div class="key plain K_COLON">			<kbd>:</kbd>			<samp>:</samp>		</div>		<div class="key plain K_QUOTE">			<kbd>"</kbd>			<samp>"</samp>		</div>		<div class="key special K_ENTER">			<kbd>Enter</kbd>		</div>		<br style="clear: left">		<div class="key special K_SHIFTL">			<kbd>Shift</kbd>		</div>		<div class="key plain K_oE2">			<kbd>\</kbd>		</div>		<div class="key plain K_Z">			<kbd>Z</kbd>			<samp>༶</samp>		</div>		<div class="key plain K_X">			<kbd>X</kbd>			<samp>྿</samp>		</div>		<div class="key plain K_C">			<kbd>C</kbd>			<samp>ྰ</samp>		</div>		<div class="key plain K_V">			<kbd>V</kbd>			<samp>ྻ</samp>		</div>		<div class="key plain K_B">			<kbd>B</kbd>			<samp>ྼ</samp>		</div>		<div class="key plain K_N empty">			<kbd>N</kbd>			<samp></samp>		</div>		<div class="key plain K_M">			<kbd>M</kbd>			<samp>ྵ</samp>		</div>		<div class="key plain K_COMMA">			<kbd>&lt;</kbd>			<samp>࿙</samp>		</div>		<div class="key plain K_PERIOD">			<kbd>&gt;</kbd>			<samp>࿚</samp>		</div>		<div class="key plain K_SLASH">			<kbd>?</kbd>			<samp>?</samp>		</div>		<div class="key special K_SHIFTR">			<kbd>Shift</kbd>		</div>		<br style="clear: left">		<div class="key special K_CTRLL">			<kbd>Ctrl</kbd>		</div>		<div class="key special K_ALTL">			<kbd>Alt</kbd>		</div>		<div class="key plain K_SPACE">			<kbd> </kbd>			<samp class='LTN'>‹no-break space›</samp>		</div>		<div class="key special K_ALTR">			<kbd>Alt</kbd>		</div>		<div class="key special K_CTRLR">			<kbd>Ctrl</kbd>		</div>	</article>
+  </section>