+++ release/el/el_nuer/HISTORY.md
@@ -1,4 +1,8 @@
-# Nuer Change History
+# Thok Nath Change History
+
+## Version 1.5 (14 Jun 2018)
+* Updated layout
+* Added cross-platform support
 
 ## 2017-08-15 1.4
 * Published to github