+++ release/g/ghana/HISTORY.md
@@ -1,5 +1,9 @@
 Ghana Keyboard Change History
 ===============================
+1.2.1 (18 Jun 2018)
+-------------------
+Add touch keyboard to package
+
 1.2 (16 Jun 2017)
 -----------------
 Removed gamma from Q and added it to G. Corrected f to j as the label for the J key.