+++ release/l/lazuri/HISTORY.md
@@ -1,6 +1,10 @@
 Lazca (Lazuri) Değişiklik Geçmişi
 ===============================
 
+1.0.1 (4 Jun 2018)
+------------------
+* Added script information to language metadata in .kps
+
 1.0 (Dec 1 2017)
 -----------------
 İlk sürüm