+++ release/s/shan/HISTORY.md
@@ -1,5 +1,12 @@
 Shan Change History
 =======================
+1.0.2 (02 Jul 2018)
+-------------------
+* Update Padauk font
+
+1.0.1 (18 Jun 2018)
+-------------------
+* Add touch keyboard to package
 
 1.0 (30 Jan 2017)
 -----------------