+++ release/sil/sil_boonkit/HISTORY.md
@@ -12,4 +12,6 @@ Boonkit Change History
  0.5.4    3-MAY-2016     subjoin a sub la to get medial la
  0.6      7-Feb-2018     Created Package and installer that installs Boonkit keyboard with fonts and Keyman Desktop 9.0
  0.6.1    7-Feb-2018     Moved to Github Keyman Repository
-
+ 0.6.2   31-May-2018     Update BCP 47 code and switch to .kvks for onscreen keyboard, minor edits to welcome and help files
+ 0.6.3   13-Jun-2018     minor edits to welcome and help files
+ 0.6.4   18-Dec-2018     Rename .png images in help files and boonkitofl.ttf font